fix mapmemory

+++ sysdeps/francium/generic/syscall.h
@@ -21,7 +21,7 @@ extern "C" {
 	//ResultCode syscall_ipc_accept(session_handle: Handle, handle_out: *mut Handle) -> ;
 	uint64_t syscall_get_process_id();
 	//ResultCode syscall_connect_to_port_handle(h: u32, handle_out: *mut Handle) -> ;
-	ResultCode syscall_map_memory(uintptr_t address, size_t length, uint32_t permission, void **addr_out);
+	ResultCode syscall_map_memory(uintptr_t address, size_t length, uint64_t permission, void **addr_out);
 	void syscall_sleep_ns(uint64_t ns);
 	uintptr_t syscall_bodge(uint32_t key, uintptr_t addr);
 	uintptr_t syscall_get_thread_id();
